pyproject.toml
[project]
name = "botsalinha"
version = "2.0.0"
description = "Discord bot specialized in Brazilian law and public contests"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core
    "agno>=2.5.5,<3.0.0",
    "discord.py>=2.6.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    # Configuration
    "pydantic>=2.10.0,<3.0.0",
    "pydantic-settings>=2.7.1,<3.0.0",
    # Database (SQLite only)
    "sqlalchemy>=2.0.35,<2.1.0",
    "alembic>=1.14.0,<2.0.0",
    # Logging
    "structlog>=24.4.0,<25.0.0",
    # Resilience
    "tenacity>=9.0.0,<10.0.0",
    # AI Model Providers (Gemini + OpenRouter via OpenAI)
    "google-genai>=1.64.0,<2.0.0",
    "openai>=1.0.0,<2.0.0",
    # RAG / Document processing
    "python-docx>=1.1.2,<2.0.0",
    "numpy>=2.0.0,<3.0.0",
    # Utilities
    "pyyaml>=6.0.2,<7.0.0",
    "aiosqlite>=0.22.1,<1.0.0",
    "cachetools>=7.0.1",
    "typer[all]>=0.24.1",
    "rich>=14.3.3",
    "questionary>=2.1.1",
    "greenlet>=3.3.2",
]

[dependency-groups]
dev = [
    # Testing framework
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.6.1",
    "pytest-env>=1.1.3",
    "pytest-timeout>=2.3.0",
    # Mocking and fixtures
    "faker>=25.1.0",
    "freezegun>=1.5.1",
    # Note: responses removed - using OpenRouter for testing instead
    # Code quality
    "ruff>=0.15.0",
    "mypy>=1.15.0",
    "types-PyYAML>=6.0.0",
    "types-cachetools>=5.3.0",
    "pre-commit>=4.1.0",
    "pip-audit>=2.8.0",
]

[project.scripts]
botsalinha = "src.main:main"
backup = "scripts.backup:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# tool.uv.dev-dependencies removed in favor of dependency-groups

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501", "SIM108"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"tests/__init__.py" = ["F401"]
"src/__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = false
warn_unused_ignores = false
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
]
